Day 1
Given an array prices, where prices[i] represents the price of a stock on the i-th day, the objective is to determine the maximum profit that can be earned by buying on one day and selling on a later day. If no profit can be made, the function should return 0.